Updated main icon.
The icon displayed in the current day is now linked to the API. In the HTML the source link was deleted since it no longer needs to be displayed as a placeholder. In the JS index the "setAttribute" function was used to replace both the image and the name of the image.